+++ plugins/org.eclipse.jem.proxy/proxyRemote/org/eclipse/jem/internal/proxy/remote/REMByteTypeBeanProxy.java
@@ -11,7 +11,7 @@ package org.eclipse.jem.internal.proxy.remote;
  *******************************************************************************/
 /*
  *  $RCSfile: REMByteTypeBeanProxy.java,v $
- *  $Revision: 1.2 $  $Date: 2005/02/15 22:56:10 $ 
+ *  $Revision: 1.3 $  $Date: 2005/05/11 22:41:24 $ 
  */
 
 
@@ -75,25 +75,25 @@ public byte byteValue() {
  * doubleValue method comment.
  */
 public double doubleValue() {
-	return (double) fByte;
+	return fByte;
 }
 /**
  * floatValue method comment.
  */
 public float floatValue() {
-	return (float) fByte;
+	return fByte;
 }
 /**
  * Return the int value
  */
 public int intValue() {
-	return (int) fByte;
+	return fByte;
 }
 /**
  * longValue method comment.
  */
 public long longValue() {
-	return (long) fByte;
+	return fByte;
 }
 /**
  * numberValue method comment.
@@ -105,7 +105,7 @@ public Number numberValue() {
  * shortValue method comment.
  */
 public short shortValue() {
-	return (short) fByte;
+	return fByte;
 }
 
 /**
