+++ plugins/org.eclipse.jem.proxy/proxyRemote/org/eclipse/jem/internal/proxy/remote/REMLongTypeBeanProxy.java
@@ -11,7 +11,7 @@ package org.eclipse.jem.internal.proxy.remote;
  *******************************************************************************/
 /*
  *  $RCSfile: REMLongTypeBeanProxy.java,v $
- *  $Revision: 1.2 $  $Date: 2005/02/15 22:56:10 $ 
+ *  $Revision: 1.3 $  $Date: 2005/05/11 22:41:24 $ 
  */
 
 
@@ -73,13 +73,13 @@ public byte byteValue() {
  * doubleValue method comment.
  */
 public double doubleValue() {
-	return (double) fLong;
+	return fLong;
 }
 /**
  * floatValue method comment.
  */
 public float floatValue() {
-	return (float) fLong;
+	return fLong;
 }
 /**
  * Return the int value
